SCB-1627 Moved CallbackContext from omega-context to omega-transaction
