+++ omega/omega-transaction/src/test/java/org/apache/servicecomb/pack/omega/transaction/CompensationMessageHandlerTest.java
@@ -28,7 +28,6 @@ import java.util.List;
 
 import org.apache.servicecomb.pack.common.EventType;
 import org.apache.servicecomb.pack.contract.grpc.ServerMeta;
-import org.apache.servicecomb.pack.omega.context.CallbackContext;
 import org.junit.Before;
 import org.junit.Test;
